refactor(rfc64): isolate graph metadata state
